fix(APIM-319): forbid any parties searchText that ends in whitespace
The previous regex for the search text was incorrect because it does not allow whitespace between
non-whitespace characters, e.g. you cannot search for "Company A".

To fix this, we could have replaced the regex with \S.
However, our code will strip any trailing whitespace and we need the searchText we send to ACBS
to have no fewer than 3 characters, so we also want the searchText to have no trailing whitespace
so the length validation is accurate.
Therefore we decided on \S$.