chore: remove redundant TODOs from APIM-492
We implemented APIM-492 to resolve an issue in the ACBS test
environment. That issue was that we were unable to create facilities
because ACBS's validation on the FacilityOverallStatus field was
incorrect. We were able to resolve this by removing the field from our
facility creation request as it is not used anyway.

We were aware that the FacilityOverallStatus field had been included in
facility creation requests from Mulesoft because the Mulesoft dev team
had found that it was required by ACBS. It was no longer required in
the ACBS test environment, but at the time of implementing APIM-492
we were unsure if it was required or not in the ACBS production
environment (which was 1 patch behind the ACBS test environment).

We now know that the FacilityOverallStatus field is not required in the
ACBS production environment (it was confirmed by FIS), so the TODOs
we left from APIM-492 (that our change might need to be reverted)
are redundant and can be removed.